# my-express-web-server/openapi.yaml
openapi: 3.0.0
info:
  title: Renewable Energy Inventory API
  description: |
    API for managing renewable energy installations, users, blogs, and transfers.
    Provides authentication, inventory management, and reporting features.
    
    Last updated: 2025-08-05
    Maintained by: Dextie
  version: 1.2.0
  contact:
    name: Support Team
    email: support@renewableenergyapp.com
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://staging-api.renewableenergyapp.com
    description: Staging server
  - url: https://api.renewableenergyapp.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Blog:
      type: object
      properties:
        _id:
          type: string
          example: "60d21b4667d0d8992e610c85"
        user:
          type: string
          example: "60d21b4667d0d8992e610c85"
        blogTitle:
          type: string
          example: "Advances in Solar Technology"
        blogSummary:
          type: string
          example: "An overview of the latest developments in solar panel efficiency."
        blogContent:
          type: string
          example: "Solar technology has seen remarkable progress in the past decade..."
        images:
          type: array
          items:
            type: string
          example: ["a1b2c3d4e5f6.webp", "f6e5d4c3b2a1.webp"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Not authorized"
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
          example: 237
        page:
          type: integer
          description: Current page number
          example: 2
        limit:
          type: integer
          description: Items per page
          example: 100
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
paths:
...
  /auth:
    post:
      summary: Login user
      description: Authenticates a user and returns access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                checkedTerms:
                  type: boolean
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '400':
          description: All fields are required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users:
    get:
      summary: Get all users
      description: Retrieves a list of all users (without password)
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: string
                description: Encrypted users data string
        '400':
          description: No users found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create a new user
      description: Creates a new user in the system
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
              required:
                - username
                - password
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "New user johndoe created"
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      summary: Update a user
      description: Updates an existing user's details
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                username:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
                active:
                  type: boolean
                password:
                  type: string
                  description: New password (optional)
                currPW:
                  type: string
                  description: Current password (required if password is being changed)
                isAdmin:
                  type: boolean
                  description: Whether the requester is an admin
                isManager:
                  type: boolean
                  description: Whether the requester is a manager
              required:
                - id
                - username
                - roles
                - active
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "johndoe updated"
        '400':
          description: Invalid data or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - password issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Delete a user
      description: Removes a user from the system
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
              required:
                - id
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: string
                example: "Username johndoe with ID 60d21b4667d0d8992e610c85 deleted"
        '400':
          description: User has assigned renewable energies or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Inventory endpoints
  /inventories:
    get:
      summary: Get all inventories
      description: Retrieves all inventory items with optional filter by solarSystemTypes
      tags:
        - Inventories
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: solarSystemTypes
          schema:
            type: string
            enum: ["Off-grid", "Grid-tied", "Hybrid"]
          description: Filter by solar system type
      responses:
        '200':
          description: Encrypted list of inventories
          content:
            application/json:
              schema:
                type: string
                description: Encrypted inventories data string
        '400':
          description: No inventories found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create new inventory
      description: Creates a new inventory item with geolocation-based duplicate detection
      tags:
        - Inventories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                type:
                  type: string
                  example: "Point"
                user:
                  type: string
                coordinates:
                  type: string
                  description: JSON string or array
                properties:
                  type: string
                  description: JSON object as string
                assessment:
                  type: string
                  description: JSON object as string
                forceCreate:
                  type: boolean
                  description: Override duplicate detection
                myUploads:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - type
                - user
                - coordinates
                - properties
      responses:
        '201':
          description: Inventory created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "New Technical assessment created"
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Potential duplicate detected
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Potential duplicate detected"
                  duplicates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inventory'
    
    patch:
      summary: Update inventory
      description: Updates an existing inventory item
      tags:
        - Inventories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                id:
                  type: string
                type:
                  type: string
                  example: "Point"
                user:
                  type: string
                coordinates:
                  type: string
                  description: JSON string or array
                properties:
                  type: string
                  description: JSON object as string
                assessment:
                  type: string
                  description: JSON object as string
                forceUpdate:
                  type: boolean
                  description: Override duplicate detection
                myUploads:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - id
                - type
                - user
                - coordinates
                - properties
      responses:
        '200':
          description: Inventory updated
          content:
            application/json:
              schema:
                type: string
                example: "'Point' updated"
        '400':
          description: Invalid data or inventory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Potential duplicate detected
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Potential duplicate detected"
                  duplicates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inventory'
    
    put:
      summary: Delete image from inventory
      description: Removes a specific image from an inventory item
      tags:
        - Inventories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                images:
                  type: number
                  description: Index of image to delete
              required:
                - id
                - images
      responses:
        '200':
          description: Image deleted from inventory
          content:
            application/json:
              schema:
                type: string
                example: "'Point' updated"
    
    delete:
      summary: Delete inventory
      description: Removes an inventory item from the system
      tags:
        - Inventories
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: array
                  items:
                    type: string
              required:
                - id
      responses:
        '200':
          description: Inventory deleted
          content:
            application/json:
              schema:
                type: string
                example: "Inventory deleted"
        '400':
          description: Invalid ID or inventory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /inventory-list:
    get:
      summary: Get paginated inventory list
      description: Returns a paginated list of inventories with filtering options
      tags:
        - Inventories
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          description: Items per page
        - in: query
          name: username
          schema:
            type: string
          description: Filter by username (for non-admin users)
        - in: query
          name: isAdmin
          schema:
            type: string
            enum: ["true", "false"]
            default: "false"
          description: Whether to apply username filter
      responses:
        '200':
          description: Paginated inventory list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inventory'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Blog endpoints
  /blogs:
    get:
      summary: Get all blogs
      description: Retrieves all blog posts
      tags:
        - Blogs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blog'
        '400':
          description: No blogs found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create new blog
      description: Creates a new blog post with images
      tags:
        - Blogs
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                user:
                  type: string
                blogTitle:
                  type: string
                blogSummary:
                  type: string
                blogContent:
                  type: string
                myUploads:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - user
                - blogTitle
                - blogSummary
                - blogContent
      responses:
        '201':
          description: Blog created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "New Technical blogContent created"
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      summary: Update blog
      description: Updates an existing blog post
      tags:
        - Blogs
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                id:
                  type: string
                user:
                  type: string
                blogTitle:
                  type: string
                blogSummary:
                  type: string
                blogContent:
                  type: string
                myUploads:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - id
                - user
                - blogTitle
                - blogSummary
                - blogContent
      responses:
        '200':
          description: Blog updated
          content:
            application/json:
              schema:
                type: string
                example: "'undefined' updated"
        '400':
          description: Invalid data or blog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Delete image from blog
      description: Removes a specific image from a blog post
      tags:
        - Blogs
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                images:
                  type: number
                  description: Index of image to delete
              required:
                - id
                - images
      responses:
        '200':
          description: Image deleted from blog
          content:
            application/json:
              schema:
                type: string
                example: "'undefined' updated"
    
    delete:
      summary: Delete blog
      description: Removes a blog post from the system
      tags:
        - Blogs
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
              required:
                - id
      responses:
        '200':
          description: Blog deleted
          content:
            application/json:
              schema:
                type: string
                example: "Blog 'undefined' with ID 60d21b4667d0d8992e610c85 deleted"
        '400':
          description: Invalid ID or blog not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /publicBlogs:
    get:
      summary: Get all public blogs
      description: Retrieves all blog posts for public viewing (encoded)
      tags:
        - Blogs
      responses:
        '200':
          description: Encoded list of blog posts
          content:
            application/json:
              schema:
                type: string
                description: Character code array as string
        '400':
          description: No blogs found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /publicInventories:
    get:
      summary: Get all public inventories
      description: Retrieves all inventory items for public viewing (encrypted)
      tags:
        - Inventories
      responses:
        '200':
          description: Encrypted list of inventories
          content:
            application/json:
              schema:
                type: string
                description: Encrypted inventories data
        '400':
          description: No inventories found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Transfer endpoints
  /transfers:
    get:
      summary: Get all transfers
      description: Retrieves all ownership transfer requests (Admin/Manager only)
      tags:
        - Transfers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Items per page
        - in: query
          name: status
          schema:
            type: string
            enum: ["pending", "approved", "rejected"]
          description: Filter by status
      responses:
        '200':
          description: List of transfers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Transfer'
                        - type: object
                          properties:
                            documents:
                              type: array
                              items:
                                type: object
                                properties:
                                  name:
                                    type: string
                                  type:
                                    type: string
                                  size:
                                    type: number
                                  hasData:
                                    type: boolean
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create transfer request
      description: Creates a new ownership transfer request
      tags:
        - Transfers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                inventoryId:
                  type: string
                newInstallerId:
                  type: string
                reason:
                  type: string
                documents:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - inventoryId
                - newInstallerId
                - reason
                - documents
      responses:
        '201':
          description: Transfer request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Transfer request created successfully"
                  transferId:
                    type: string
                    example: "60d21b4667d0d8992e610c85"
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Inventory or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []